refactor: reimplement filtered image in detect image content layer
