+++ app/src/main/java/com/madrid/movio/app.kt
@@ -1,7 +1,6 @@
 package com.madrid.movio
 
 
-import com.madrid.presentation.screens.searchScreen.SearchViewModel
 import com.madrid.data.dataSource.local.SearchLocalDataSource
 import com.madrid.data.dataSource.remote.search.SearchRemoteSourceImpl
 import com.madrid.data.repositories.SearchLocalSource
@@ -15,6 +14,7 @@ import com.madrid.domain.usecase.searchUseCase.MediaUseCase
 import com.madrid.domain.usecase.searchUseCase.PreferredMediaUseCase
 import com.madrid.domain.usecase.searchUseCase.RecentSearchUseCase
 import com.madrid.domain.usecase.searchUseCase.TrendingMediaUseCase
+import com.madrid.presentation.screens.searchScreen.SearchViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
@@ -36,7 +36,6 @@ val app = module {
             get(),
         )
     }
-
 
     //domain
     single { ArtistUseCase(get()) }
