--- presentation/src/main/java/com/madrid/presentation/composables/filteredImage/FilteredImage.kt
@@ -1,75 +1,0 @@
-package com.madrid.presentation.composables.filteredImage
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.size
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.blur
-import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.graphics.DefaultAlpha
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.unit.dp
-import org.koin.androidx.compose.koinViewModel
-
-
-@Composable
-fun FilteredImage(
-    imageUrl: String,
-    contentDescription: String?,
-    modifier: Modifier = Modifier,
-    alignment: Alignment = Alignment.Center,
-    contentScale: ContentScale = ContentScale.Fit,
-    alpha: Float = DefaultAlpha,
-    colorFilter: ColorFilter? = null,
-    viewModel: FilteredImageViewModel = koinViewModel()
-) {
-    val state by viewModel.state.collectAsState()
-
-    LaunchedEffect(imageUrl) {
-        viewModel.loadAndAnalyze(imageUrl)
-    }
-
-    FilteredImageContent(
-        state,
-        contentDescription = contentDescription,
-        modifier = modifier,
-        alignment = alignment,
-        contentScale = contentScale,
-        alpha = alpha,
-        colorFilter = colorFilter
-    )
-}
-
-
-@Composable
-fun FilteredImageContent(
-    state: FilteredImageUiState,
-    contentDescription: String?,
-    modifier: Modifier = Modifier,
-    alignment: Alignment = Alignment.Center,
-    contentScale: ContentScale = ContentScale.Fit,
-    alpha: Float = DefaultAlpha,
-    colorFilter: ColorFilter? = null
-) {
-    state.bitmap?.let {
-        Image(
-            bitmap = it.asImageBitmap(),
-            contentDescription = contentDescription,
-            modifier = modifier
-                .size(300.dp)
-                .then(
-                    if (state.isNSFW) Modifier.blur(16.dp)
-                    else Modifier
-                ),
-            contentScale = contentScale,
-            alignment = alignment,
-            alpha = alpha,
-            colorFilter = colorFilter
-        )
-    }
-}
