--- presentation/src/main/java/com/madrid/presentation/composables/filteredImage/FilteredImageUiState.kt
@@ -1,9 +1,0 @@
-package com.madrid.presentation.composables.filteredImage
-
-import android.graphics.Bitmap
-
-data class FilteredImageUiState(
-    val bitmap: Bitmap? = null,
-    val isNSFW: Boolean = false,
-    val error: String? = null
-)
